Getting price history query price parsing sorted.
Got the price history model objects working and parsing correctly now. Also fixed the date/time parsing
so the latest price can be identified (added tests to make sure it parsed and calculated values right).
Also kept a copy of the prices returned from CREST for the "Vespa II Blueprint" that is getting used in
the test cases.